Backend/Backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
drf-spectacular>=0.27
django-ratelimit>=4.1.0
gunicorn>=23.0
psycopg[binary]>=3.2  # PostgreSQL adapter
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0


